Complete mySql benchmarking test scripts
