+++ db/mySql/index.js
@@ -1,0 +1,12 @@
+const mysql = require('mysql');
+
+const connection = mysql.createConnection({
+  host: 'localhost',
+  user: 'root',
+  // password : 'secret',
+  database: 'sdc_reviews',
+});
+
+connection.connect();
+
+module.exports = connection;
